moved types to separate file
